requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib
bcrypt
pyjwt
email_validator
alembic
